requirements.txt
PyYAML~=5.1.2
numpy~=1.17.2
pandas~=0.25.1
torch~=1.4.0
tqdm~=4.32.1
scikit-image~=0.16.1
matplotlib~=3.1.1
Pillow~=6.2.0
sklearn~=0.0
scikit-learn~=0.21.3
torchvision~=0.5.0
setuptools~=41.4.0
tensorboardX~=1.6
pretrainedmodels~=0.7.4
tensorflow~=1.13.1
tensorflow-gpu~=1.13.1
Keras~=2.3.1
scipy~=1.4.1